feat(provisioning): update github workflow

+++ packages/gramine/scripts/start.sh
@@ -7,27 +7,23 @@ FOLDER="./dist"
 URL_ZIP="https://github.com/hack-a-chain-software/bazk/releases/download/bazk-poc/dist.zip"
 
 # Docker command to run
-DOCKER_COMMAND="sudo docker run --env .env --rm --device /dev/sgx_enclave --device /dev/sgx_provision -v $(pwd)/dist:/dist -it gramineproject/gramine"
+DOCKER_COMMAND="sudo docker run --env-file .env --rm --device /dev/sgx_enclave --device /dev/sgx_provision -v $(pwd)/dist:/dist -it gramineproject/gramine"
 
-# Function to download and extract ZIP file
-update_folder() {
-    echo "Downloading and updating folder $FOLDER..."
+# Checks if the folder exists
+if [ -d "$FOLDER" ]; then
+    echo "Folder $FOLDER found. Running Docker..."
+else
+    echo "Folder $FOLDER not found. Downloading and extracting the file..."
 
     # Downloads the ZIP file
     curl -L -o dist.zip $URL_ZIP
-
-    # Removes the existing folder if it exists
-    [ -d "$FOLDER" ] && rm -rf "$FOLDER"
 
     # Extracts the ZIP file
     unzip dist.zip -d ./
 
     # Removes the ZIP file after extraction
     rm dist.zip
-}
-
-# Always update the folder with the latest version
-update_folder
+fi
 
 # Executes the Docker command
 $DOCKER_COMMAND
